+++ Runtime/ARKitLoader.cs
@@ -10,7 +10,7 @@ namespace UnityEngine.XR.ARKit
         private static List<XRCameraSubsystemDescriptor> s_CameraSubsystemDescriptors = new List<XRCameraSubsystemDescriptor>();
         private static List<XRDepthSubsystemDescriptor> s_DepthSubsystemDescriptors = new List<XRDepthSubsystemDescriptor>();
         private static List<XRPlaneSubsystemDescriptor> s_PlaneSubsystemDescriptors = new List<XRPlaneSubsystemDescriptor>();
-        private static List<XRReferencePointSubsystemDescriptor> s_ReferencePointSubsystemDescriptors = new List<XRReferencePointSubsystemDescriptor>();
+        private static List<XRAnchorSubsystemDescriptor> s_AnchorSubsystemDescriptors = new List<XRAnchorSubsystemDescriptor>();
         private static List<XRRaycastSubsystemDescriptor> s_RaycastSubsystemDescriptors = new List<XRRaycastSubsystemDescriptor>();
         private static List<XRHumanBodySubsystemDescriptor> s_HumanBodySubsystemDescriptors = new List<XRHumanBodySubsystemDescriptor>();
         private static List<XREnvironmentProbeSubsystemDescriptor> s_EnvironmentProbeSubsystemDescriptors = new List<XREnvironmentProbeSubsystemDescriptor>();
@@ -51,11 +51,11 @@ namespace UnityEngine.XR.ARKit
             }
         }
 
-        public XRReferencePointSubsystem referencePointSubsystem
+        public XRAnchorSubsystem anchorSubsystem
         {
             get
             {
-                return GetLoadedSubsystem<XRReferencePointSubsystem>();
+                return GetLoadedSubsystem<XRAnchorSubsystem>();
             }
         }
 
@@ -122,7 +122,7 @@ namespace UnityEngine.XR.ARKit
             CreateSubsystem<XRCameraSubsystemDescriptor, XRCameraSubsystem>(s_CameraSubsystemDescriptors, "ARKit-Camera");
             CreateSubsystem<XRDepthSubsystemDescriptor, XRDepthSubsystem>(s_DepthSubsystemDescriptors, "ARKit-Depth");
             CreateSubsystem<XRPlaneSubsystemDescriptor, XRPlaneSubsystem>(s_PlaneSubsystemDescriptors, "ARKit-Plane");
-            CreateSubsystem<XRReferencePointSubsystemDescriptor, XRReferencePointSubsystem>(s_ReferencePointSubsystemDescriptors, "ARKit-ReferencePoint");
+            CreateSubsystem<XRAnchorSubsystemDescriptor, XRAnchorSubsystem>(s_AnchorSubsystemDescriptors, "ARKit-Anchor");
             CreateSubsystem<XRRaycastSubsystemDescriptor, XRRaycastSubsystem>(s_RaycastSubsystemDescriptors, "ARKit-Raycast");
             CreateSubsystem<XRHumanBodySubsystemDescriptor, XRHumanBodySubsystem>(s_HumanBodySubsystemDescriptors, "ARKit-HumanBody");
             CreateSubsystem<XREnvironmentProbeSubsystemDescriptor, XREnvironmentProbeSubsystem>(s_EnvironmentProbeSubsystemDescriptors, "ARKit-EnvironmentProbe");
@@ -153,7 +153,7 @@ namespace UnityEngine.XR.ARKit
                 StartSubsystem<XRCameraSubsystem>();
                 StartSubsystem<XRDepthSubsystem>();
                 StartSubsystem<XRPlaneSubsystem>();
-                StartSubsystem<XRReferencePointSubsystem>();
+                StartSubsystem<XRAnchorSubsystem>();
                 StartSubsystem<XRRaycastSubsystem>();
                 StartSubsystem<XRHumanBodySubsystem>();
                 StartSubsystem<XREnvironmentProbeSubsystem>();
@@ -174,7 +174,7 @@ namespace UnityEngine.XR.ARKit
                 StopSubsystem<XRCameraSubsystem>();
                 StopSubsystem<XRDepthSubsystem>();
                 StopSubsystem<XRPlaneSubsystem>();
-                StopSubsystem<XRReferencePointSubsystem>();
+                StopSubsystem<XRAnchorSubsystem>();
                 StopSubsystem<XRRaycastSubsystem>();
                 StopSubsystem<XRHumanBodySubsystem>();
                 StopSubsystem<XREnvironmentProbeSubsystem>();
@@ -193,7 +193,7 @@ namespace UnityEngine.XR.ARKit
             DestroySubsystem<XRCameraSubsystem>();
             DestroySubsystem<XRDepthSubsystem>();
             DestroySubsystem<XRPlaneSubsystem>();
-            DestroySubsystem<XRReferencePointSubsystem>();
+            DestroySubsystem<XRAnchorSubsystem>();
             DestroySubsystem<XRRaycastSubsystem>();
             DestroySubsystem<XRHumanBodySubsystem>();
             DestroySubsystem<XREnvironmentProbeSubsystem>();